refactor: tempArticle과 관련된 타입끼리 분리 및 적용
