+++ frontend/src/hooks/tempArticle/useGetTempArticles.tsx
@@ -1,8 +1,8 @@
 import { AxiosError } from 'axios';
 import { useQuery } from 'react-query';
 
-import { TotalTempArticleResponseType } from '@/api/article/articleType';
 import { getTempArticles } from '@/api/tempArticle/tempArticle';
+import { TotalTempArticleResponseType } from '@/api/tempArticle/tempArticleType';
 import { ErrorMessage } from '@/constants/ErrorMessage';
 import useThrowCustomError from '@/hooks/common/useThrowCustomError';
